Cleanup closures.
