+++ iml-action-client/Cargo.toml
@@ -9,7 +9,6 @@ futures = "0.3"
 iml-action-runner = { path = "../iml-services/iml-action-runner", version = "0.2" }
 iml-manager-env = { path = "../iml-manager-env", version = "0.2" }
 iml-wire-types = { path = "../iml-wire-types", version = "0.2" }
-lazy_static = "1.4.0"
 reqwest = { version = "0.10", features = ["default-tls", "native-tls", "json", "stream"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
