+++ iml-action-client/src/lib.rs
@@ -10,30 +10,10 @@ use futures::{
     TryFutureExt,
 };
 use iml_action_runner::ActionType;
-use iml_manager_env::{get_action_runner_connect, get_action_runner_socket};
+use iml_manager_env::get_action_runner_connect;
 use iml_wire_types::{Action, ActionId, ActionName, Fqdn};
-use lazy_static::lazy_static;
-use std::io::BufRead;
 use thiserror::Error;
 use uuid::Uuid;
-
-lazy_static! {
-    static ref RUNNING_IN_DOCKER: bool = {
-        match std::fs::File::open("/proc/self/cgroup") {
-            Err(_) => false,
-            Ok(file) => {
-                let reader = std::io::BufReader::new(file);
-                reader
-                    .lines()
-                    .filter(|l| {
-                        l.as_ref().unwrap_or(&"".to_string()).split(':').nth(1) == Some("docker")
-                    })
-                    .next()
-                    .is_some()
-            }
-        }
-    };
-}
 
 #[derive(Error, Debug)]
 pub enum ImlActionClientError {
@@ -49,12 +29,7 @@ pub fn invoke_rust_agent(
     args: impl serde::Serialize,
 ) -> (oneshot::Sender<()>, impl Future<Output = Result<serde_json::Value, ImlActionClientError>>) {
     let request_id = Uuid::new_v4().to_hyphenated().to_string();
-    let conn = if *RUNNING_IN_DOCKER {
-        get_action_runner_connect()
-    } else {
-        get_action_runner_socket()
-    };
-    let conn2 = conn.clone();
+    let conn = get_action_runner_connect();
 
     let action = ActionType::Remote((
         host.clone().into(),
@@ -69,38 +44,34 @@ pub fn invoke_rust_agent(
 
     let (p, c) = oneshot::channel::<()>();
 
-    let post = async move {
-        client
-            .post(&conn)
-            .json(&action)
-            .send()
-            .await?
-            .json()
-            .err_into()
-            .await
-    }.boxed();
+    let post = client
+        .post(&conn)
+        .json(&action)
+        .send()
+        .err_into()
+        .and_then(|resp| resp.json().err_into())
+        .boxed();
 
-    let fut = future::select(c, post)
-        .then(|either| async move {
-            if let Either::Right((_,_)) = either {
+    let fut = async move {
+        let either = future::select(c, post).await;
+
+        match either {
+            Either::Left((x, f)) => {
+                if let Err(_) = x {
+                    return f.await;
+                };
                 let cancel = ActionType::Remote((
                     host.into(),
                     Action::ActionCancel {
                         id: ActionId(request_id),
-                    }
+                    },
                 ));
-                let client = reqwest::Client::new();
-                let _rc = client
-                    .post(&conn2)
-                    .json(&cancel)
-                    .send()
-                    .await;
+                let _rc = client.post(&conn).json(&cancel).send().await;
+                Err(ImlActionClientError::CancelledRequest)
             }
-            either
-        }).map(|either| match either {
-            Either::Left((_, b)) => { drop(b); Err(ImlActionClientError::CancelledRequest) }
-            Either::Right((a, _)) => a,
-        });
+            Either::Right((x, _)) => x,
+        }
+    };
 
     (p, fut)
 }
