+++ iml-manager-env/Cargo.toml
@@ -5,4 +5,5 @@ authors = ["IML Team <iml@whamcloud.com>"]
 edition = "2018"
 
 [dependencies]
+lazy_static = "1.4.0"
 url = "2.1"
